EEPROM y Correccion INTERFACES

+++ Firmware_v2/paterf/inc/rs485.h
@@ -8,6 +8,6 @@
 uint8_t RS485_Flag();
 void    RS485_GetBuffer     (uint8_t[SIZE_BUFFER] buffer);
 void    RS485_SetState      (estado_t state);
-void    RS485_SetBaudRate   (uint8_t baudRate);
+void    RS485_SetBaudRate   (uint32_t baudRate);
 void    RS485_Write         (uint8_t[SIZE_BUFFER] data);
 #endif
