+++ Firmware_v2/paterf/inc/usb.h
@@ -7,6 +7,6 @@
 uint8_t USB_Flag();
 void    USB_GetBuffer     (uint8_t[SIZE_BUFFER] buffer);
 void    USB_SetState      (estado_t state);
-void    USB_SetBaudRate   (uint8_t baudRate);
+void    USB_SetBaudRate   (uint32_t baudRate);
 void    USB_Write         (uint8_t[SIZE_BUFFER] data);
 #endif
